refactor tutorial pages
