+++ web-app/src/containers/Tutorial/Completed/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import * as T from 'typings'
 import { css, jsx } from '@emotion/core'
-import Button from '../../components/Button'
+import Button from '../../../components/Button'
 
 const styles = {
   page: {
